Fix regression from #511 (#515)
* M1233650 Part 1

* M1233650 Part 2+3

* M1233650 Part 4
